+++ src/code/reducers/helpers/gems-helper.js
